feat(OH-116): 환자 추가 페이지 공통 레이아웃으로 구현
